rework part 1 to prep for part 2 start work on part 2
There seems to be a major flaw in my logic.
It works on the smaller test input,
but it is nearly an entire order of magnitude off on the second.
I am projecting that it will be multiple orders of magnitude off on the actual input.
(possibly 2 or 3)

Looking at the results, I believe that my grouping and counting on those groups
are the culprits in the miscalculation.

In long chains of 1-gaps, I cut off at each of the 3-gaps.
There could be many miscounted permutations bridged over those gaps.